backend/requirements.txt
asgiref==3.3.1
astroid==2.5.1
Django==3.1.7
django-cors-headers==3.10.0
django-debug-toolbar==3.2
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
environs==1.2.0
isort==5.7.0
lazy-object-proxy==1.5.2
marshmallow==3.14.0
mccabe==0.6.1
PyJWT==2.3.0
pylint==2.7.2
pylint-django==2.4.2
pylint-plugin-utils==0.6
pytz==2021.1
read-env==1.1.0
sqlparse==0.4.1
toml==0.10.2
wrapt==1.12.1